// src/AdsetManagementWeb/src/app/api/models/VehicleListResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { VehicleResponse } from './VehicleResponse';
export type VehicleListResponse = {
    data?: Array<VehicleResponse> | null;
    totalItems?: number;
    page?: number;
    pageSize?: number;
    totalPages?: number;
};